requirements.txt
asgiref==3.10.0
Django==5.2.7
django-environ==0.12.0
et_xmlfile==2.0.0
gunicorn==23.0.0
openpyxl==3.1.5
packaging==25.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
